+++ packages/engine_umbrella/apps/engine/lib/vehicle.ex
@@ -59,8 +59,6 @@ defmodule Vehicle do
     {:reply, updated_vehicle, updated_vehicle}
   end
 
-  def generate_id, do: "pmv-" <> Engine.Utils.generate_id()
-
   def publish(data, routing_key),
     do:
       @rmq.publish(
@@ -69,11 +67,10 @@ defmodule Vehicle do
         routing_key
       )
 
-  def make(%{start_address: start_address} = vehicle_info) do
+  def make(%{id: _id, start_address: start_address} = vehicle_info) do
     vehicle_fields =
       vehicle_info
       |> Map.put_new(:end_address, start_address)
-      |> Map.put(:id, generate_id())
       |> Map.put_new(:capacity, %{volume: 15, weight: 700})
       |> Map.update(:metadata, nil, &Jason.encode!/1)
 
@@ -108,7 +105,7 @@ defmodule Vehicle do
     vehicle
   end
 
-  def update(%{id: "pmv-" <> _ = id} = vehicle_update) do
+  def update(%{id: "pmt-" <> _ = id} = vehicle_update) do
     updated_vehicle =
       get(id)
       |> Map.merge(vehicle_update)
@@ -132,13 +129,13 @@ defmodule Vehicle do
     true
   end
 
-  def apply_offer_accepted("pmv-" <> _ = id, offer) do
+  def apply_offer_accepted("pmt-" <> _ = id, offer) do
     GenServer.call(via_tuple(id), {:apply_offer_accepted, offer})
     |> Map.put(:current_route, get_route_from_activities(offer.activities))
     |> publish("new_instructions")
   end
 
-  def apply_vehicle_to_state(%Vehicle{id: "pmv-" <> _ = id} = vehicle) do
+  def apply_vehicle_to_state(%Vehicle{id: "pmt-" <> _ = id} = vehicle) do
     GenServer.start_link(
       __MODULE__,
       vehicle,
@@ -152,12 +149,12 @@ defmodule Vehicle do
     vehicle
   end
 
-  def delete("pmv-" <> _ = id) do
+  def delete("pmt-" <> _ = id) do
     ES.add_event(%VehicleDeleted{id: id})
     apply_delete_to_state(id)
   end
 
-  def apply_delete_to_state("pmv-" <> _ = id) do
+  def apply_delete_to_state("pmt-" <> _ = id) do
     Engine.VehicleStore.delete_vehicle(id)
     GenServer.stop(via_tuple(id))
 
@@ -166,7 +163,7 @@ defmodule Vehicle do
 
   defp via_tuple(id), do: {:via, :gproc, {:n, :l, {:vehicle_id, id}}}
 
-  def get("pmv-" <> _ = id), do: GenServer.call(via_tuple(id), :get)
+  def get("pmt-" <> _ = id), do: GenServer.call(via_tuple(id), :get)
 
   defp get_route_from_activities(activities),
     do:
@@ -174,7 +171,7 @@ defmodule Vehicle do
       |> Enum.map(fn %{address: address} -> address end)
       |> Osrm.route()
 
-  def offer(%Vehicle{id: "pmv-" <> _ = id, activities: activities, booking_ids: booking_ids}) do
+  def offer(%Vehicle{id: "pmt-" <> _ = id, activities: activities, booking_ids: booking_ids}) do
     offer = %{
       booking_ids: booking_ids,
       activities: activities
@@ -196,7 +193,7 @@ defmodule Vehicle do
     end
   end
 
-  def send_offer(offer, "pmv-" <> _ = vehicle_id) do
+  def send_offer(offer, "pmt-" <> _ = vehicle_id) do
     Logger.debug("offer to vehicle #{vehicle_id}")
 
     RMQRPCWorker.call(
