+++ packages/engine_umbrella/apps/message_generator/lib/transport_generator.ex
@@ -1,5 +1,6 @@
 defmodule MessageGenerator.TransportGenerator do
   alias MessageGenerator.Address
+  alias MessageGenerator.Id
 
   @stockholm %{lat: 59.3414072, lon: 18.0470482}
   @gothenburg %{lat: 57.7009147, lon: 11.7537571}
@@ -8,7 +9,7 @@ defmodule MessageGenerator.TransportGenerator do
   defp default_metadata() do
     %{
       profile: "Generated transport #{Enum.random(1..100)}",
-      driver: %{ contact: "4676000#{Enum.random(1000..9999)}"}
+      driver: %{contact: "4676000#{Enum.random(1000..9999)}"}
     }
   end
 
@@ -16,6 +17,7 @@ defmodule MessageGenerator.TransportGenerator do
     default_meta = default_metadata()
 
     properties
+    |> Map.put_new(:id, Id.generate_transport_id())
     |> put_new_transport_addresses_from_city(:ljusdal)
     |> Map.update(:metadata, default_meta, &Map.merge(default_meta, &1))
   end
